Blog - extract filtering logic from CategoryTab.tsx in blog.tsx
