+++ src/pages/blog.tsx
@@ -1,6 +1,6 @@
 import Head from 'next/head'
 import Link from 'next/link'
-import React, { useState } from 'react'
+import React, { useMemo, useState } from 'react'
 
 import Button from '../components/Button'
 import CategoryTab from '../components/CategoryTab'
@@ -9,6 +9,7 @@ import publicRuntimeConfig from '../config/env.config'
 import { getAllWhitePapers, getCategories, getIdeasPageContent, getIdeasByPage } from '../Services/wordpressService'
 import { Category, IdeasDesc, IdeasPageContent } from '../types/IdeasContent'
 import WhitePaper from '../types/WhitePaper'
+import IdeasCard from '../components/IdeasCard'
 import clsx from 'clsx'
 import ContactSection from '../components/ContactSection'
 import Particles from '../static/images/particles-3.svg'
@@ -33,10 +34,57 @@ function compareIdeasByDate(idea1: IdeasDesc, idea2: IdeasDesc) {
 
 function Ideas({ ideasDescription, whitePapers, categories, content }: Props) {
   const [sortedIdeas, setSortedIdeas] = useState(ideasDescription.sort(compareIdeasByDate))
-  const [filter, setFilter] = useState('')
+  const [searchFilter, setSearchFilter] = useState('')
   const [isOpenenedModal, setOpenModal] = useState(false)
   const [isError, setIsError] = useState(true)
   const [isSubmited, setIsSubmited] = useState(false)
+  const [searchTimeout, setSearchTimeout] = useState(null)
+  const [categoryFilter, setCategoryFilter] = useState('All')
+  const [loadedPages, setLoadedPages] = useState(1)
+  const [isLoading, setIsLoading] = useState(false)
+
+  const fakeIdea1 = {
+    id: -1,
+    slug: '',
+    title: '',
+    category: '',
+    categories: [''],
+    tags: [],
+    descriptionText: '',
+    date: '',
+    image: content.ideasimg1,
+  }
+
+  const fakeIdea2 = {
+    id: -2,
+    slug: '',
+    title: '',
+    category: '',
+    categories: [''],
+    tags: [],
+    descriptionText: '',
+    date: '',
+    image: content.ideasimg2,
+  }
+
+  function getBgColor(category: string) {
+    if (category === 'Agaetis' || category === 'Evènements') {
+      return 'bg-orange'
+    } else if (category === 'Stratégie SI') {
+      return 'bg-blue'
+    } else if (category === 'Data') {
+      return 'bg-teal'
+    } else if (category === 'Service Design') {
+      return 'bg-light-pink'
+    } else if (category === 'Technologie') {
+      return 'bg-yellow'
+    } else if (category === 'Agilité') {
+      return 'bg-light-purple'
+    } else if (category === 'Business Hacking') {
+      return 'bg-red-lighter'
+    }
+    return 'bg-grey'
+  }
 
   function handleOpenModal(error: boolean) {
     setIsError(error)
@@ -57,8 +105,13 @@ function Ideas({ ideasDescription, whitePapers, categories, content }: Props) {
     }
   }
 
-  const [loadedPages, setLoadedPages] = useState(1)
-  const [isLoading, setIsLoading] = useState(false)
+  function handleFilterChange(category: string) {
+    setCategoryFilter(category)
+  }
+
+  function createMarkup(content: string) {
+    return { __html: content }
+  }
 
   async function handleFetchMoreIdeas() {
     setIsLoading(true)
@@ -94,6 +147,39 @@ function Ideas({ ideasDescription, whitePapers, categories, content }: Props) {
     setIsLoading(false)
   }
 
+  const filteredIdeas = useMemo(() => {
+    const ideas = sortedIdeas.filter(
+      (idea) =>
+        !idea.categories.includes('White-paper') &&
+        !idea.categories.includes('Jobs') &&
+        idea.title.toLocaleLowerCase().includes(searchFilter.toLocaleLowerCase()) &&
+        (categoryFilter === 'All' || idea.categories.includes(categoryFilter) || idea.categories.includes(''))
+    )
+
+    if (content.ideasimg1 && !ideas.find((idea: IdeasDesc) => idea.id === -1)) {
+      ideas.splice(0, 0, fakeIdea1)
+    }
+    if (content.ideasimg2 && !ideas.find((idea: IdeasDesc) => idea.id === -2)) {
+      ideas.splice(7, 0, fakeIdea2)
+    }
+    return ideas.map((idea) => {
+      return (
+        <div key={idea.id} className={clsx(ideas.length > 2 ? 'sm:w-1/3' : 'sm:w-1/2', ' p-2')}>
+          <IdeasCard
+            className={clsx(
+              { 'shadow-xl hidden sm:block': idea.id < 0 },
+              { [getBgColor(idea.categories.filter((category) => !category.includes('_offer-'))[0])]: idea.id > 0 }
+            )}
+            {...idea}
+            categories={idea.categories.filter((category) => !category.includes('_offer-'))}
+          >
+            <p dangerouslySetInnerHTML={createMarkup(idea.descriptionText)} />
+          </IdeasCard>
+        </div>
+      )
+    })
+  }, [categoryFilter, fakeIdea1, fakeIdea2])
+
   return (
     <>
       <Head>
@@ -121,24 +207,22 @@ function Ideas({ ideasDescription, whitePapers, categories, content }: Props) {
               </div>
               <SearchInput // To change
                 handleChange={(value: string) => {
-                  setFilter(value)
+                  if (searchTimeout) clearTimeout(searchTimeout)
+                  setSearchTimeout(setTimeout(() => setSearchFilter(value), 500))
                 }}
               ></SearchInput>
             </div>
 
             <CategoryTab
-              ideasC={sortedIdeas.filter(
-                (idea) =>
-                  !idea.categories.includes('White-paper') &&
-                  !idea.categories.includes('Jobs') &&
-                  idea.title.toLocaleLowerCase().includes(filter.toLocaleLowerCase())
-              )}
               categories={categories.filter(
                 (category) => category.categoryName !== 'Jobs' && category.categoryName !== 'White-paper'
               )}
-              ideasImg1={content.ideasimg1}
-              ideasImg2={content.ideasimg2}
+              categoryFilter={categoryFilter}
+              handleFilterChange={handleFilterChange}
             />
+            <div className="flex flex-col md:max-w-lg sm:flex-row justify-center flex-wrap mt-2 md:p-8 mx-auto">
+              {filteredIdeas}
+            </div>
             <Button
               className={clsx(
                 'flex flex-row justify-center uppercase rounded-full bg-orange text-xss py-2 px-6 text-white font-semibold mx-auto see-more',
